# apps/server/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: SiteCRM API
  version: 0.2.0
  description: |
    Primary REST API for SiteCRM.  Handles authentication, lead management,
    outreach, notifications, site generation, and deployments.

    **Authentication:** Most endpoints require a Bearer JWT.  Obtain a token
    via `POST /api/auth/login` or `POST /api/auth/register` and pass it as
    `Authorization: Bearer <token>` on subsequent requests.

    **Common error responses (apply to every operation unless noted):**
    - `400 Bad Request` — request body or query failed validation
      (`additionalProperties: false` rejects unknown fields).
    - `401 Unauthorized` — missing, invalid, or expired token on a protected route.
    - `503 Service Unavailable` — the database is not configured/reachable.

    All errors use the shared `Error` schema. Reusable definitions are under
    `components/responses` (`BadRequest`, `Unauthorized`, `NotFound`,
    `ServiceUnavailable`).
  contact:
    name: Riverfront Labs
servers:
  - url: http://api.localhost
    description: Local development (Traefik)
  - url: http://localhost:3000
    description: Direct local development
...
paths:
...
  /api/notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications (newest first)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: unread
          in: query
          description: When true, returns only unread notifications.
          schema: { type: boolean }
      responses:
        '200':
          description: Paginated notification list (includes total unread count)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNotifications'
  /api/notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all unread notifications as read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Number of notifications marked read
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
  /api/notifications/{notificationId}/read:
    parameters:
      - name: notificationId
        in: path
        required: true
        schema: { type: string, format: uuid }

    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notifications/stream:
    get:
      operationId: notificationStream
      tags: [notifications]
      summary: SSE stream for real-time notifications
      description: |
        Server-Sent Events stream. The browser `EventSource` API cannot send
        custom headers, so authenticate via `?token=<jwt>` instead of
        `Authorization: Bearer`.

        Event types pushed to the client:
        - `connected` — sent immediately on open
        - `notification` — a new {@link Notification} arrived
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: query
          description: JWT — alternative to the Authorization header for EventSource clients.
          schema: { type: string }
      responses:
        '200':
          description: SSE stream (Content-Type text/event-stream)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Reusable error responses — reference from operations via
  # `$ref: '#/components/responses/<Name>'`. See the info description for when
  # each applies across the API.
...
  schemas:
...
    NotificationType:
      type: string
      enum:
        - dm_received
        - email_reply
        - lead_status_change
        - deployment_complete
        - build_complete
        - job_complete
        - job_failed
    Notification:
      type: object
      required: [id, userId, type, title, body, read, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/NotificationType'
        title:
          type: string
        body:
          type: string
        data:
          type: ['object', 'null']
          description: Arbitrary JSON payload — e.g. `{ leadId }` for deep-link navigation.
        read:
          type: boolean
        readAt:
          type: ['string', 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
    PaginatedNotifications:
      type: object
      required: [data, total, page, pageSize, unreadCount]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        unreadCount:
          type: integer
          description: Total unread count across all pages — use for badge display.

    # ── Shared ────────────────────────────────────────────────────────────────
    Error:
      type: object
      required: [statusCode, error, message]
      description: Standard error envelope — matches Fastify's default error shape.
      properties:
        statusCode:
          type: integer
        error:
          type: string
          description: HTTP reason phrase, e.g. `Not Found`
        message:
          type: string
          description: Human-readable explanation of what went wrong